[2607] Improve the AttachmentFactory API design
