+++ docusaurus/docs/Android/04-compose/02-general-customization/02-attachment-factory.mdx
@@ -5,11 +5,11 @@ The `AttachmentFactory` class allows you to build your own attachments to displa
 ```kotlin
 public open class AttachmentFactory @ExperimentalStreamChatApi constructor(
     public val canHandle: (attachments: List<Attachment>) -> Boolean,
-    public val previewContent: @Composable (
+    public val previewContent: (@Composable (
         modifier: Modifier,
         attachments: List<Attachment>,
         onAttachmentRemoved: (Attachment) -> Unit,
-    ) -> Unit,
+    ) -> Unit)? = null,
     public val content: @Composable (
         modifier: Modifier,
         attachmentState: AttachmentState,
