+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/attachments/factory/GiphyAttachmentFactory.kt
@@ -2,7 +2,6 @@ package io.getstream.chat.android.compose.ui.attachments.factory
 
 import androidx.compose.runtime.Composable
 import com.getstream.sdk.chat.model.ModelType
-import io.getstream.chat.android.compose.ui.attachments.AttachmentFactory
 import io.getstream.chat.android.compose.ui.attachments.content.GiphyAttachmentContent
 
 /**
@@ -14,7 +13,6 @@ import io.getstream.chat.android.compose.ui.attachments.content.GiphyAttachmentC
 @Suppress("FunctionName")
 public fun GiphyAttachmentFactory(): AttachmentFactory = AttachmentFactory(
     canHandle = { attachments -> attachments.any { it.type == ModelType.attach_giphy } },
-    previewContent = @Composable { _, _, _ -> },
     content = @Composable { modifier, state ->
         GiphyAttachmentContent(
             modifier = modifier,
