+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/attachments/factory/ImageAttachmentFactory.kt
@@ -1,7 +1,6 @@
 package io.getstream.chat.android.compose.ui.attachments.factory
 
 import androidx.compose.runtime.Composable
-import io.getstream.chat.android.compose.ui.attachments.AttachmentFactory
 import io.getstream.chat.android.compose.ui.attachments.content.ImageAttachmentContent
 import io.getstream.chat.android.compose.ui.attachments.content.ImageAttachmentPreviewContent
 import io.getstream.chat.android.compose.ui.util.isMedia
