+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/attachments/factory/LinkAttachmentFactory.kt
@@ -2,7 +2,6 @@ package io.getstream.chat.android.compose.ui.attachments.factory
 
 import androidx.compose.runtime.Composable
 import com.getstream.sdk.chat.model.ModelType
-import io.getstream.chat.android.compose.ui.attachments.AttachmentFactory
 import io.getstream.chat.android.compose.ui.attachments.content.LinkAttachmentContent
 import io.getstream.chat.android.compose.ui.util.hasLink
 
@@ -19,7 +18,6 @@ public fun LinkAttachmentFactory(
     linkDescriptionMaxLines: Int,
 ): AttachmentFactory = AttachmentFactory(
     canHandle = { links -> links.any { it.hasLink() && it.type != ModelType.attach_giphy } },
-    previewContent = @Composable { _, _, _ -> },
     content = @Composable { modifier, state ->
         LinkAttachmentContent(
             modifier = modifier,
